+++ ext/IntegralsCubaExt.jl
@@ -1,135 +1,7 @@
-module IntegralsCuba
+module IntegralsCubaExt
 
 using Integrals, Cuba
 import Integrals: transformation_if_inf, scale_x, scale_x!
-
-abstract type AbstractCubaAlgorithm <: SciMLBase.AbstractIntegralAlgorithm end
-"""
-    CubaVegas()
-
-Multidimensional adaptive Monte Carlo integration from Cuba.jl.
-Importance sampling is used to reduce variance.
-
-## References
-
-@article{lepage1978new,
-title={A new algorithm for adaptive multidimensional integration},
-author={Lepage, G Peter},
-journal={Journal of Computational Physics},
-volume={27},
-number={2},
-pages={192--203},
-year={1978},
-publisher={Elsevier}
-}
-"""
-struct CubaVegas <: AbstractCubaAlgorithm
-    flags::Int
-    seed::Int
-    minevals::Int
-    nstart::Int
-    nincrease::Int
-    gridno::Int
-end
-"""
-    CubaSUAVE()
-
-Multidimensional adaptive Monte Carlo integration from Cuba.jl.
-Suave stands for subregion-adaptive VEGAS.
-Importance sampling and subdivision are thus used to reduce variance.
-
-## References
-
-@article{hahn2005cuba,
-title={Cuba—a library for multidimensional numerical integration},
-author={Hahn, Thomas},
-journal={Computer Physics Communications},
-volume={168},
-number={2},
-pages={78--95},
-year={2005},
-publisher={Elsevier}
-}
-"""
-struct CubaSUAVE{R} <: AbstractCubaAlgorithm where {R <: Real}
-    flags::Int
-    seed::Int
-    minevals::Int
-    nnew::Int
-    nmin::Int
-    flatness::R
-end
-"""
-    CubaDivonne()
-
-Multidimensional adaptive Monte Carlo integration from Cuba.jl.
-Stratified sampling is used to reduce variance.
-
-## References
-
-@article{friedman1981nested,
-title={A nested partitioning procedure for numerical multiple integration},
-author={Friedman, Jerome H and Wright, Margaret H},
-journal={ACM Transactions on Mathematical Software (TOMS)},
-volume={7},
-number={1},
-pages={76--92},
-year={1981},
-publisher={ACM New York, NY, USA}
-}
-"""
-struct CubaDivonne{R1, R2, R3} <:
-       AbstractCubaAlgorithm where {R1 <: Real, R2 <: Real, R3 <: Real}
-    flags::Int
-    seed::Int
-    minevals::Int
-    key1::Int
-    key2::Int
-    key3::Int
-    maxpass::Int
-    border::R1
-    maxchisq::R2
-    mindeviation::R3
-end
-"""
-    CubaCuhre()
-
-Multidimensional h-adaptive integration from Cuba.jl.
-
-## References
-
-@article{berntsen1991adaptive,
-title={An adaptive algorithm for the approximate calculation of multiple integrals},
-author={Berntsen, Jarle and Espelid, Terje O and Genz, Alan},
-journal={ACM Transactions on Mathematical Software (TOMS)},
-volume={17},
-number={4},
-pages={437--451},
-year={1991},
-publisher={ACM New York, NY, USA}
-}
-"""
-struct CubaCuhre <: AbstractCubaAlgorithm
-    flags::Int
-    minevals::Int
-    key::Int
-end
-
-function CubaVegas(; flags = 0, seed = 0, minevals = 0, nstart = 1000, nincrease = 500,
-        gridno = 0)
-    CubaVegas(flags, seed, minevals, nstart, nincrease, gridno)
-end
-function CubaSUAVE(; flags = 0, seed = 0, minevals = 0, nnew = 1000, nmin = 2,
-        flatness = 25.0)
-    CubaSUAVE(flags, seed, minevals, nnew, nmin, flatness)
-end
-function CubaDivonne(; flags = 0, seed = 0, minevals = 0,
-        key1 = 47, key2 = 1, key3 = 1, maxpass = 5, border = 0.0,
-        maxchisq = 10.0, mindeviation = 0.25)
-    CubaDivonne(flags, seed, minevals, key1, key2, key3, maxpass, border, maxchisq,
-        mindeviation)
-end
-CubaCuhre(; flags = 0, minevals = 0, key = 0) = CubaCuhre(flags, minevals, key)
 
 function Integrals.__solvebp_call(prob::IntegralProblem, alg::AbstractCubaAlgorithm,
         sensealg,
@@ -242,7 +114,5 @@ function Integrals.__solvebp_call(prob::IntegralProblem, alg::AbstractCubaAlgori
     SciMLBase.build_solution(prob, alg, val, out.error,
         chi = out.probability, retcode = ReturnCode.Success)
 end
-
-export CubaVegas, CubaSUAVE, CubaDivonne, CubaCuhre
 
 end
