--- lib/IntegralsCuba/Project.toml
@@ -1,19 +1,0 @@
-name = "IntegralsCuba"
-uuid = "e00cd5f1-6337-4131-8b37-28b2fe4cd6cb"
-authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
-version = "0.3.0"
-
-[deps]
-Cuba = "8a292aeb-7a57-582c-b821-06e4c11590b1"
-Integrals = "de52edbc-65ea-441a-8357-d3a637375a31"
-
-[compat]
-Cuba = "2"
-Integrals = "3"
-julia = "1.6"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
