--- lib/IntegralsCuba/test/runtests.jl
@@ -1,1 +1,0 @@
-
