--- lib/IntegralsCubature/Project.toml
@@ -1,19 +1,0 @@
-name = "IntegralsCubature"
-uuid = "c31f79ba-6e32-46d4-a52f-182a8ac42a54"
-authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
-version = "0.2.3"
-
-[deps]
-Cubature = "667455a9-e2ce-5579-9412-b964f529a492"
-Integrals = "de52edbc-65ea-441a-8357-d3a637375a31"
-
-[compat]
-Cubature = "1"
-Integrals = "3"
-julia = "1.6"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
